+++ CHANGES.md
@@ -1,3 +1,19 @@
+0.15.0.0
+--------
+
+- Data.Packed.Foreign (additional FFI helpers)
+
+- NFData instance of Matrix
+
+- Unidimensional root finding
+
+- In Numeric.LinearAlgebra.Util:
+    pairwise2D, rowOuters, null1, null1sym, size, unitary, mt, (¦), (?), (¿)
+
+- diagBlock
+
+- meanCov moved to Container
+
 0.14.1.0
 --------